Fix LibreNMS Website Logo SVG missing i sometimes (#119)
* Update LibreNMS-Logo.svg

Corrected an issue with the "i" being invisible in certain scenarios on certain (Chromium version 77) browsers, see https://community.librenms.org/t/librenms-svg-logo-issue-in-some-zooms/9640/4. The issue seems to be that "i" was a separate clip-path from the "L bre".

* Update LibreNMS-Logo-Black.svg

Corrected an issue with the "i" being invisible in certain scenarios on certain (Chromium version 77) browsers, see https://community.librenms.org/t/librenms-svg-logo-issue-in-some-zooms/9640/4. The issue seems to be that "i" was a separate clip-path from the "L bre".